odilia_cache::event_handlers

Trait EventHandler

pub trait EventHandler {
    // Required method
    fn handle_event<D: CacheDriver + Send>(
        self,
        cache: &mut Cache<D>,
    ) -> impl Future<Output = Result<CacheItem, OdiliaError>> + Send;
}

Implemented for all events which modify items in the cache in some way.

Required Methods

fn handle_event<D: CacheDriver + Send>( self, cache: &mut Cache<D>, ) -> impl Future<Output = Result<CacheItem, OdiliaError>> + Send

How does this event affect the cache. It might remove links between cache items, remove them completely, add new ones, interact the the CacheDriver directly. The only thing that is required by implementers of this function is that the cache remains in some generally consistent state; for example, if you are removing a child from its parent, you must remove the relationship on both sides.

Generally this should be implemented by those that understand the invariants of the accessibility tree.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.