Function odilia_cache::accessible_to_cache_item

pub async fn accessible_to_cache_item(
    accessible: &AccessibleProxy<'_>,
    cache: Weak<Cache>
) -> OdiliaResult<CacheItem>

Convert an atspi::accessible::AccessibleProxy into a crate::CacheItem. This runs a bunch of long-awaiting code and can take quite some time; use this sparingly. This takes most properties and some function calls through the AccessibleProxy structure and generates a new CacheItem, which will be written to cache before being sent back.

Errors

Will return an Err(_) variant when:

The cache parameter does not reference an active cache once the Weak is upgraded to an Option<Arc<_>>.

Any of the function calls on the accessible fail.

Any (String, OwnedObjectPath) -> AccessiblePrimitive conversions fail. This should never happen, but technically it is possible.