Trait odilia_cache::AccessibleExt

pub trait AccessibleExt {
    type Error: Error;

    // Required methods
    fn get_application_ext<'a>(
        &self
    ) -> impl Future<Output = Result<AccessibleProxy<'a>, Self::Error>> + Send
       where Self: Sized;
    fn get_parent_ext<'a>(
        &self
    ) -> impl Future<Output = Result<AccessibleProxy<'a>, Self::Error>> + Send
       where Self: Sized;
    fn get_children_ext<'a>(
        &self
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_siblings<'a>(
        &self
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_children_indexes<'a>(
        &self
    ) -> impl Future<Output = Result<Vec<i32>, Self::Error>> + Send;
    fn get_siblings_before<'a>(
        &self
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_siblings_after<'a>(
        &self
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_children_from_caret<'a>(
        &self,
        after: bool
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_next<'a>(
        &self,
        role: Role,
        backward: bool
    ) -> impl Future<Output = Result<Option<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn edges<'a>(
        &self,
        backward: Option<bool>
    ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
       where Self: Sized;
    fn get_relation_set_ext<'a>(
        &self
    ) -> impl Future<Output = Result<HashMap<RelationType, Vec<AccessibleProxy<'a>>>, Self::Error>> + Send
       where Self: Sized;
    fn match_(
        &self,
        role: Role
    ) -> impl Future<Output = Result<bool, OdiliaError>> + Send;
}

Required Associated Types

type Error: Error

Required Methods

fn get_application_ext<'a>( &self ) -> impl Future<Output = Result<AccessibleProxy<'a>, Self::Error>> + Send
where Self: Sized,

fn get_parent_ext<'a>( &self ) -> impl Future<Output = Result<AccessibleProxy<'a>, Self::Error>> + Send
where Self: Sized,

fn get_children_ext<'a>( &self ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn get_siblings<'a>( &self ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn get_children_indexes<'a>( &self ) -> impl Future<Output = Result<Vec<i32>, Self::Error>> + Send

fn get_siblings_before<'a>( &self ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn get_siblings_after<'a>( &self ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn get_children_from_caret<'a>( &self, after: bool ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn get_next<'a>( &self, role: Role, backward: bool ) -> impl Future<Output = Result<Option<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

fn edges<'a>( &self, backward: Option<bool> ) -> impl Future<Output = Result<Vec<AccessibleProxy<'a>>, Self::Error>> + Send
where Self: Sized,

Get all edges for a given accessible object. This means: all children, siblings, and parent, in that order. If a direction is specified, then it will only get the appicable matching siblings/children. This also checks if the element supports the text interface, and then checks if the caret position is contained within the string, if it is, then children are also handled by direction.

fn get_relation_set_ext<'a>( &self ) -> impl Future<Output = Result<HashMap<RelationType, Vec<AccessibleProxy<'a>>>, Self::Error>> + Send
where Self: Sized,

fn match_( &self, role: Role ) -> impl Future<Output = Result<bool, OdiliaError>> + Send

Object Safety

This trait is not object safe.