odilia_cache

Trait CacheDriver

pub trait CacheDriver {
    // Required methods
    fn lookup_external(
        &self,
        key: &CacheKey,
    ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send;
    fn lookup_bulk(
        &self,
        key: &CacheKey,
    ) -> impl Future<Output = OdiliaResult<Vec<CacheItem>>> + Send;
    fn lookup_relations(
        &self,
        key: &CacheKey,
        ty: RelationType,
    ) -> impl Future<Output = OdiliaResult<Vec<CacheKey>>> + Send;
    fn lookup_from_cache_item(
        &self,
        cache_item: CacheItem,
    ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send;
    fn lookup_from_legacy_cache_item(
        &self,
        cache_item: LegacyCacheItem,
    ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send;
}

A method of performing I/O side-effects outside the cache itself. This is made an explicit trait such that we can either:

Call out to DBus (production), or

Use a fixed set of items (testing).

Panic when called.

Feel free to implement your own at your convenience.

Required Methods

fn lookup_external( &self, key: &CacheKey, ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send

Lookup a given CacheKey that was not found in the cache..

fn lookup_bulk( &self, key: &CacheKey, ) -> impl Future<Output = OdiliaResult<Vec<CacheItem>>> + Send

Bulk query an application based on the [CacheKey.sender] field.

fn lookup_relations( &self, key: &CacheKey, ty: RelationType, ) -> impl Future<Output = OdiliaResult<Vec<CacheKey>>> + Send

A seperate method from [lookup_external] for getting relation sets. This is separate because it can have rather large results and should only be called when absolutely necessary.

fn lookup_from_cache_item( &self, cache_item: CacheItem, ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send

A separate method from [lookup_external] for converting an atspi::CacheItem into a CacheItem. This will call out to DBus for the remaining details.

fn lookup_from_legacy_cache_item( &self, cache_item: LegacyCacheItem, ) -> impl Future<Output = OdiliaResult<CacheItem>> + Send

A separate method from [lookup_external] for converting an atspi::LegacyCacheItem into a CacheItem. This will call out to DBus for the remaining details.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.