odilia_cache

Trait Convertable

pub trait Convertable<'a> {
    type Error: Error;

    // Required methods
    fn to_accessible(
        &self,
    ) -> impl Future<Output = Result<AccessibleProxy<'_>, Self::Error>> + Send;
    fn to_action(
        &self,
    ) -> impl Future<Output = Result<ActionProxy<'_>, Self::Error>> + Send;
    fn to_application(
        &self,
    ) -> impl Future<Output = Result<ApplicationProxy<'_>, Self::Error>> + Send;
    fn to_collection(
        &self,
    ) -> impl Future<Output = Result<CollectionProxy<'_>, Self::Error>> + Send;
    fn to_component(
        &self,
    ) -> impl Future<Output = Result<ComponentProxy<'_>, Self::Error>> + Send;
    fn to_document(
        &self,
    ) -> impl Future<Output = Result<DocumentProxy<'_>, Self::Error>> + Send;
    fn to_hypertext(
        &self,
    ) -> impl Future<Output = Result<HypertextProxy<'_>, Self::Error>> + Send;
    fn to_hyperlink(
        &self,
    ) -> impl Future<Output = Result<HyperlinkProxy<'_>, Self::Error>> + Send;
    fn to_image(
        &self,
    ) -> impl Future<Output = Result<ImageProxy<'_>, Self::Error>> + Send;
    fn to_selection(
        &self,
    ) -> impl Future<Output = Result<SelectionProxy<'_>, Self::Error>> + Send;
    fn to_table(
        &self,
    ) -> impl Future<Output = Result<TableProxy<'_>, Self::Error>> + Send;
    fn to_table_cell(
        &self,
    ) -> impl Future<Output = Result<TableCellProxy<'_>, Self::Error>> + Send;
    fn to_text(
        &self,
    ) -> impl Future<Output = Result<TextProxy<'_>, Self::Error>> + Send;
    fn to_editable_text(
        &self,
    ) -> impl Future<Output = Result<EditableTextProxy<'_>, Self::Error>> + Send;
    fn to_value(
        &self,
    ) -> impl Future<Output = Result<ValueProxy<'_>, Self::Error>> + Send;

}

Required Associated Types

type Error: Error

Required Methods

fn to_accessible( &self, ) -> impl Future<Output = Result<AccessibleProxy<'_>, Self::Error>> + Send

Creates an AccessibleProxy from the existing accessible item.

Errors

This may fail based on the implementation of. Generally, it fails if the accessible item does not implement to accessible interface. This shouldn’t be possible, but this function may fail for other reasons. For example, to convert a zbus::Proxy into a AccessibleProxy, it may fail to create the new atspi::proxy::accessible::AccessibleProxy.

fn to_action( &self, ) -> impl Future<Output = Result<ActionProxy<'_>, Self::Error>> + Send

Creates an ActionProxy from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to action interface.

fn to_application( &self, ) -> impl Future<Output = Result<ApplicationProxy<'_>, Self::Error>> + Send

Creates an ApplicationProxy from the existing accessible item.

Errors

This may fail based on the implementation. Generally, it fails if the accessible item does not implement to application interface.

fn to_collection( &self, ) -> impl Future<Output = Result<CollectionProxy<'_>, Self::Error>> + Send

Creates an CollectionProxy from the existing accessible item.

Errors

This may fail based on the implementation. it fails if the accessible item does not implement to collection interface.

fn to_component( &self, ) -> impl Future<Output = Result<ComponentProxy<'_>, Self::Error>> + Send

Creates an ComponentProxy from the existing accessible item.

Errors

This may fail based on the implementation. it fails if the accessible item does not implement to component interface.

fn to_document( &self, ) -> impl Future<Output = Result<DocumentProxy<'_>, Self::Error>> + Send

fn to_hypertext( &self, ) -> impl Future<Output = Result<HypertextProxy<'_>, Self::Error>> + Send

fn to_hyperlink( &self, ) -> impl Future<Output = Result<HyperlinkProxy<'_>, Self::Error>> + Send

fn to_image( &self, ) -> impl Future<Output = Result<ImageProxy<'_>, Self::Error>> + Send

fn to_selection( &self, ) -> impl Future<Output = Result<SelectionProxy<'_>, Self::Error>> + Send

fn to_table( &self, ) -> impl Future<Output = Result<TableProxy<'_>, Self::Error>> + Send

fn to_table_cell( &self, ) -> impl Future<Output = Result<TableCellProxy<'_>, Self::Error>> + Send

fn to_text( &self, ) -> impl Future<Output = Result<TextProxy<'_>, Self::Error>> + Send

fn to_editable_text( &self, ) -> impl Future<Output = Result<EditableTextProxy<'_>, Self::Error>> + Send

fn to_value( &self, ) -> impl Future<Output = Result<ValueProxy<'_>, Self::Error>> + Send

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.