Crate odilia_cache

Modules

accessible_ext 🔒

convertable 🔒

Structs

Cache

An internal cache used within Odilia.

CacheEffectDBus

CacheEffectFixedList

CacheItem

A struct representing an accessible. To get any information from the cache other than the stored information like role, interfaces, and states, you will need to instantiate an atspi_proxies::accessible::AccessibleProxy or other *Proxy type from atspi to query further info.

RelationSet

Enums

Link

Traits

AccessibleExt

AllText 🔒

CacheExt

CacheSideEffect

A method of performing I/O side-effects outside the cache itself. This is made an explicit trait such that we can either:

Convertable

Functions

accessible_to_cache_item

Convert an atspi_proxies::accessible::AccessibleProxy into a crate::CacheItem. This runs a bunch of long-awaiting code and can take quite some time; use this sparingly. This takes most properties and some function calls through the AccessibleProxy structure and generates a new CacheItem, which will be written to cache before being sent back.

Type Aliases

CacheKey 🔒

IdLookupTable 🔒

ThreadSafeCache 🔒