Crate odilia_cache

Modules

accessible_ext 🔒

convertable 🔒

event_handlers 🔒

relation_set 🔒

Structs

Cache

An internal cache used within Odilia.

CacheActor

A method of interacting with the cache. All requests on the cache side are processed in synchronous FIFO order.

CacheItem

A struct representing an accessible. To get any information from the cache other than the stored information like role, interfaces, and states, you will need to instantiate an atspi::proxy::accessible::AccessibleProxy or other *Proxy type from atspi to query further info.

Children

ControlledBy

ControllerFor

DescribedBy

DescriptionFor

Details

DetailsFor

EmbeddedBy

Embeds

ErrorFor

ErrorMessage

FlowsFrom

FlowsTo

Item

LabelFor

LabelledBy

MemberOf

NewCache 🔒

NodeChildOf

NodeParentOf

Parent

ParentWindowOf

PopupFor

RelationSet

Relations

SubwindowOf

Enums

CacheRequest

CacheResponse

Traits

AccessibleExt

AllText 🔒

CacheDriver

A method of performing I/O side-effects outside the cache itself. This is made an explicit trait such that we can either:

CacheExt

ConstRelationType

Convertable

EventHandler

Implemented for all events which modify items in the cache in some way.

Functions

accessible_to_cache_item

Convert an atspi::proxy::accessible::AccessibleProxy into a crate::CacheItem. This runs a bunch of long-awaiting code and can take quite some time; use this sparingly. This takes most properties and some function calls through the AccessibleProxy structure and generates a new CacheItem, which will be written to cache before being sent back.

cache_handler_task

or_cancel 🔒

Type Aliases

ActorRecv

ActorRequest

ActorSend

CacheKey