odilia_cache

Struct Cache

pub struct Cache<D: CacheDriver> {
    pub(crate) tree: NewCache,
    pub driver: D,
}

An internal cache used within Odilia.

This contains (mostly) all accessibles in the entire accessibility tree, and they are referenced by their IDs. If you are having issues with incorrect or invalid accessibles trying to be accessed, this is code is probably the issue.

Fields
tree: NewCachedriver: D
Implementations

impl<D: CacheDriver + Send> Cache<D>

pub(crate) async fn handle_event( &mut self, ev: Event, ) -> Result<CacheItem, OdiliaError>

pub(crate) async fn request( &mut self, req: CacheRequest, ) -> Result<CacheResponse, OdiliaError>

pub fn tree( &self, ) -> &HashMap<String, HashMap<String, CacheItem, FxBuildHasher>, FxBuildHasher>

impl<D: CacheDriver> Cache<D>

pub fn new(driver: D) -> Self

create a new, fresh cache

pub fn remove(&mut self, id: &CacheKey) -> Option<CacheItem>

Remove a single cache item. This function can not fail.

pub fn get(&self, id: &CacheKey) -> Option<CacheItem>

Get a single item from the cache.

This will allow you to get the item without holding any locks to it, at the cost of (1) a clone and (2) no guarantees that the data is kept up-to-date.

pub fn get_all(&self, ids: &[CacheKey]) -> Vec<Option<CacheItem>>

get a many items from the cache; this only creates one read handle (note that this will copy all data you would like to access)

pub fn remove_all(&mut self, ids: &[CacheKey])

Bulk remove all ids in the cache; this only refreshes the cache after removing all items.

pub fn add(&mut self, ci: CacheItem) -> CacheItem

pub fn add_all(&mut self, cis: Vec<CacheItem>) -> Vec<CacheItem>

pub(crate) async fn prefetch_app( &mut self, key: &CacheKey, ) -> OdiliaResult<CacheItem>

pub(crate) async fn get_or_create_all( &mut self, keys: Vec<CacheKey>, ) -> OdiliaResult<Vec<CacheItem>>

pub async fn modify_if_not_new<F>( &mut self, key: &AccessiblePrimitive, f: F, ) -> OdiliaResult<CacheItem>
where F: FnOnce(&mut CacheItem),

Modify the given item with closure [F] if it was already contained in the cache. Otherwise, fetch a new item over the CacheDriver.

Errors

See: [get_or_create]

pub(crate) async fn get_or_create( &mut self, key: &AccessiblePrimitive, ) -> OdiliaResult<CacheItem>

Get a single item from the cache (note that this copies some integers to a new struct). If the CacheItem is not found, create one, add it to the cache, and return it.

Errors

The function will return an error if:

The accessible can not be turned into an AccessiblePrimitive. This should never happen, but is technically possible.

The Self::add function fails.

The accessible_to_cache_item function fails.

Panics

This function technically has a .expect() which could panic. But we gaurs against this.

pub(crate) async fn get_or_create_from_cache_item( &mut self, ci: CacheItem, ) -> OdiliaResult<CacheItem>

Same as [get_or_create] but starts with an initial atspi::CacheItem.

Errors

The function will return an error if:

The accessible can not be turned into an AccessiblePrimitive. This should never happen, but is technically possible.

The Self::add function fails.

The accessible_to_cache_item function fails.

Panics

This function technically has a .expect() which could panic. But we gaurs against this.

pub(crate) async fn get_or_create_from_legacy_cache_item( &mut self, ci: LegacyCacheItem, ) -> OdiliaResult<CacheItem>

Same as [get_or_create] but starts with an initial atspi::LegacyCacheItem.

Errors

The function will return an error if:

The accessible can not be turned into an AccessiblePrimitive. This should never happen, but is technically possible.

The Self::add function fails.

The accessible_to_cache_item function fails.

Panics

This function technically has a .expect() which could panic. But we gaurs against this.

pub fn clear(&mut self)