Struct odilia_cache::CacheItem

pub struct CacheItem {
    pub object: AccessiblePrimitive,
    pub app: AccessiblePrimitive,
    pub parent: CacheRef,
    pub index: i32,
    pub children_num: i32,
    pub interfaces: InterfaceSet,
    pub role: Role,
    pub states: StateSet,
    pub text: String,
    pub children: Vec<CacheRef>,
    pub cache: Weak<Cache>,
}

A struct representing an accessible. To get any information from the cache other than the stored information like role, interfaces, and states, you will need to instantiate an atspi_proxies::accessible::AccessibleProxy or other *Proxy type from atspi to query further info.

Fields
object: AccessiblePrimitiveapp: AccessiblePrimitiveparent: CacheRefindex: i32children_num: i32interfaces: InterfaceSetrole: Rolestates: StateSettext: Stringchildren: Vec<CacheRef>cache: Weak<Cache>
Implementations

impl CacheItem

pub fn parent_ref(&mut self) -> OdiliaResult<Arc<RwLock<CacheItem>>>

Return a reference to a parent. This is much cheaper than getting the parent element outright via Self::parent.

Errors

This method will return a CacheError::NoItem if no item is found within the cache.

pub async fn from_atspi_event<'a, T: GenericEvent<'a>>( event: &T, cache: Weak<Cache>, connection: &Connection ) -> OdiliaResult<Self>

Creates a CacheItem from an atspi::Event type.

Errors

This can fail under three possible conditions:

We are unable to convert information from the event into an AccessiblePrimitive hashmap key. This should never happen.

We are unable to convert the AccessiblePrimitive to an atspi_proxies::accessible::AccessibleProxy.

The accessible_to_cache_item function fails for any reason. This also shouldn’t happen.

pub async fn from_atspi_cache_item( atspi_cache_item: CacheItem, cache: Weak<Cache>, connection: &Connection ) -> OdiliaResult<Self>

Convert an atspi::CacheItem into a crate::CacheItem. This requires calls to DBus, which is quite expensive. Beware calling this too often.

Errors

This function can fail under the following conditions:

The atspi::CacheItem can not be turned into a crate::AccessiblePrimitive. This should never happen.

The crate::AccessiblePrimitive can not be turned into a atspi_proxies::accessible::AccessibleProxy. This should never happen.

Getting children from the AccessibleProxy fails. This should never happen.

The only time these can fail is if the item is removed on the application side before the conversion to AccessibleProxy.

pub fn get_children(&self) -> OdiliaResult<Vec<Self>>

Get a Vec of children with the same type as Self.

Errors

Will return an Err variant if self.cache does not reference an active cache. This should never happen, but it is technically possible.

Any children keys’ values are not found in the cache itself.